+++ libs/features/mas/web-components/build.mjs
@@ -69,15 +69,6 @@ Promise.all([
         plugins: [rewriteImports()],
         external: ['lit'],
     }),
-    build({
-      entryPoints: ['./src/cdt/countdown-timer.js'],
-      bundle: true,
-      minify: true,
-      outfile: `${outfolder}/countdown-timer.js`,
-      format: 'esm',
-      plugins: [rewriteImports()],
-      external: ['lit'],
-    }),
     buildLitComponent('merch-icon'),
     buildLitComponent('merch-quantity-select'),
     buildLitComponent('merch-secure-transaction'),
