+++ libs/features/mas/web-components/src/index.js
@@ -1,3 +1,2 @@
 export * from './merch-card.js';
 export * from './merch-card-collection.js';
-export * from './cdt/countdown-timer.js';
